Initial Commit - v 0.0.1-SNAPSHOT
Add base application setup with TestContainers integration

This commit introduces the base setup for a Spring Boot Kotlin application. The application features integration with TestContainers not just for testing, but also for development purposes. Added build configuration details and initial classes for a Person module,which serves as an example of application behavior. Database schema initialization is handled by Flyway, with migration scripts included. Demonstrations of both development and testing uses of TestContainers are provided in the test classes.